managing unauthorize access

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use App\Http\Requests\StorePost;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -68,6 +78,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if ($post->user_id !== auth()->user()->id) {
+            return redirect('/posts')->with('error', 'Unauthorize Access');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -82,6 +97,11 @@ class PostsController extends Controller
     {
         // Create post
         $post = Post::find($id);
+
+        if ($post->user_id !== auth()->user()->id) {
+            return redirect('/posts')->with('error', 'Unauthorize Access');
+        }
+        
         $post->title = $request->title;
         $post->body = $request->body;
         $post->save();
@@ -97,7 +117,13 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
+        
         $post = Post::find($id);
+
+        if ($post->user_id !== auth()->user()->id) {
+            return redirect('/posts')->with('error', 'Unauthorize Access');
+        }
+        
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Deleted');
